+++ Oppgave2/text1.txt
@@ -1,0 +1,4 @@
+Tester linjeskift.
+Og en til ...
+Og en til ...
+Og en til ...
