+++ Oppgave2/text2.txt
@@ -1,0 +1,4 @@
+Tester linjeskift.
+Og en til ...
+Og en til ...
+Og en til ...
