--- files/text2.txt
@@ -1,4 +1,0 @@
-Tester linjeskift.
-Og en til ...
-Og en til ...
-Og en til ...
